+++ src/meta/raft-store/src/leveled_store/leveled_map/mod.rs
@@ -13,9 +13,6 @@
 // limitations under the License.
 
 use std::collections::BTreeMap;
-use std::io;
-use std::io::Error;
-use std::ops::RangeBounds;
 use std::sync::Arc;
 use std::sync::Mutex;
 
@@ -26,33 +23,18 @@ use databend_common_meta_types::raft_types::StoredMembership;
 use databend_common_meta_types::snapshot_db::DB;
 use databend_common_meta_types::sys_data::SysData;
 use databend_common_meta_types::Node;
-use futures_util::StreamExt;
 use log::info;
-use log::warn;
 use map_api::mvcc;
-use map_api::mvcc::ViewKey;
-use map_api::mvcc::ViewValue;
-use map_api::util;
-use map_api::IOResultStream;
-use map_api::MapKey;
 use seq_marked::InternalSeq;
-use seq_marked::SeqMarked;
-use stream_more::KMerge;
-use stream_more::StreamMore;
 
 use crate::leveled_store::immutable::Immutable;
 use crate::leveled_store::immutable_levels::ImmutableLevels;
-use crate::leveled_store::level::GetTable;
-use crate::leveled_store::level::Level;
 use crate::leveled_store::leveled_map::applier_acquirer::WriterPermit;
 use crate::leveled_store::leveled_map::compacting_data::CompactingData;
 use crate::leveled_store::leveled_map::immutable_data::ImmutableData;
 use crate::leveled_store::leveled_map::leveled_map_data::LeveledMapData;
-use crate::leveled_store::map_api::MapKeyDecode;
-use crate::leveled_store::map_api::MapKeyEncode;
 use crate::leveled_store::snapshot::MvccSnapshot;
 use crate::leveled_store::snapshot::StateMachineSnapshot;
-use crate::leveled_store::value_convert::ValueConvert;
 use crate::leveled_store::view::StateMachineView;
 
 #[cfg(test)]
@@ -64,8 +46,10 @@ pub mod compactor;
 pub mod compactor_acquirer;
 pub mod immutable_data;
 mod impl_commit;
-mod impl_snapshot_get;
-mod impl_snapshot_range;
+mod impl_scoped_seq_bounded_get;
+mod impl_scoped_seq_bounded_range;
+mod impl_seq_bounded_get;
+mod impl_seq_bounded_range;
 pub mod leveled_map_data;
 
 #[cfg(test)]
@@ -98,94 +82,6 @@ impl Default for LeveledMap {
         Self {
             data: Arc::new(Mutex::new(LeveledMapData::default())),
         }
-    }
-}
-
-// TODO: test it
-#[async_trait::async_trait]
-impl<K> mvcc::ScopedSeqBoundedGet<K, K::V> for LeveledMap
-where
-    K: MapKey,
-    K: ViewKey,
-    K: MapKeyEncode + MapKeyDecode,
-    K::V: ViewValue,
-    SeqMarked<K::V>: ValueConvert<SeqMarked>,
-    Level: GetTable<K, K::V>,
-    Immutable: mvcc::ScopedSeqBoundedGet<K, K::V>,
-{
-    async fn get(&self, key: K, snapshot_seq: u64) -> Result<SeqMarked<K::V>, io::Error> {
-        let immutable = {
-            let inner = self.data.lock().unwrap();
-            let got = inner
-                .writable
-                .get_table()
-                .get(key.clone(), snapshot_seq)
-                .cloned();
-            if !got.is_not_found() {
-                return Ok(got);
-            }
-
-            inner.immutable.clone()
-        };
-
-        immutable.get(key, snapshot_seq).await
-    }
-}
-
-// TODO: test it
-#[async_trait::async_trait]
-impl<K> mvcc::ScopedSeqBoundedRange<K, K::V> for LeveledMap
-where
-    K: MapKey,
-    K: ViewKey,
-    K: MapKeyEncode + MapKeyDecode,
-    K::V: ViewValue,
-    SeqMarked<K::V>: ValueConvert<SeqMarked>,
-    Level: GetTable<K, K::V>,
-    Immutable: mvcc::ScopedSeqBoundedRange<K, K::V>,
-{
-    async fn range<R>(
-        &self,
-        range: R,
-        snapshot_seq: u64,
-    ) -> Result<IOResultStream<(K, SeqMarked<K::V>)>, Error>
-    where
-        R: RangeBounds<K> + Send + Sync + Clone + 'static,
-    {
-        let mut kmerge = KMerge::by(util::by_key_seq);
-
-        // writable level
-
-        let (vec, immutable) = {
-            let inner = self.data.lock().unwrap();
-            let it = inner
-                .writable
-                .get_table()
-                .range(range.clone(), snapshot_seq);
-            let vec = it.map(|(k, v)| (k.clone(), v.cloned())).collect::<Vec<_>>();
-
-            (vec, inner.immutable.clone())
-        };
-
-        if vec.len() > 1000 {
-            warn!(
-                "Level.writable::range(start={:?}, end={:?}) returns big range of len={}",
-                range.start_bound(),
-                range.end_bound(),
-                vec.len()
-            );
-        }
-
-        let strm = futures::stream::iter(vec).map(Ok).boxed();
-        kmerge = kmerge.merge(strm);
-
-        let strm = immutable.range(range, snapshot_seq).await?;
-        kmerge = kmerge.merge(strm);
-
-        // Merge entries with the same key, keep the one with larger internal-seq
-        let coalesce = kmerge.coalesce(util::merge_kv_results);
-
-        Ok(coalesce.boxed())
     }
 }
